Add them

+++ app/Http/Controllers/Admin/MovieController.php
@@ -91,12 +91,12 @@ class MovieController extends Controller
         ];
         if($request->file('image'))
         {
-            Storage::delete($ticket->image);
+            if($ticket->image) Storage::delete($ticket->image);
             $data['image'] = Storage::put(Movies::$image_store_path, $request->file('image'));
         }
         if($request->file('base_ticket_image'))
         {
-            Storage::delete($ticket->base_ticket_image);
+            if($ticket->base_ticket_image) Storage::delete($ticket->base_ticket_image);
             $data['base_ticket_image'] = Storage::put(Movies::$image_store_path, $request->file('base_ticket_image'));
         }
